adjust errortrace output to not lose some expressions
By using a combination of the original and expanded programs,
errortrace attempts to provide a better printed form of an
expression; that support was added in a6494ffc8c (I think).
When `disarm` has any effect, however, a new syntax object is
allocated, so the `eq?`-based comparisons of that tracking
fail.

A complete solution probably involves registering an equivalence
between original and disarmed forms. For now, it's at least
better to keep the expanded code's form than end up with #f.