confine errortrace to initial source terms
We don't want to confine errortrace instrumentation to
`syntax-original?` terms, because we want to show expressions
that are from macros defined in instrumented modules. But
we don't want to instrument the detailed expansion of
`for`, for example; that creates much more run-time overhead
for no expected debugging benefit. The new approach here
is to add a syntax property to code given to errortrace ---
much like `syntax-original?`, but without a connection to
macro-introduction scopes.

At the stacktrace level, there's a new `stacktrace/filter@`
unit (etc.) to make the `should-annotate?` test configurable.
It's the `errortrace/errortrace-lib` layer that adds the
new property.

The top level can partially expand a form before errortrace sees it.
In that case, errortrace may instrument some expressions that are the
result of macro expansion fro ma non-annotated module. We accept this
approximation, because existing uses of errortrace can synthesize
forms with a source location but not `syntax-original?`ness, and
we want errortrace to keep working for those cases.