include call to functions like `list` in a stack trace
Errortrace annotation formerly skipped a wrapper around calls to
functions like `list`, which can never fail. That's no good if the
argument to the function can fail, though.

If the arguments also can never fail, then the wraper is unnecessary,
but the compiler should be able to detect those cases and eliminate
the wrapper in many cases.

Eli noticed this problem when observing DrRacket's stacktrace arrows
for the program that is in the new test.